+++ system/platform-core/src/test/java/org/platformlambda/automation/RestEndpointTest.java
@@ -49,7 +49,7 @@ class RestEndpointTest extends TestBase {
     private static final SimpleXmlParser xml = new SimpleXmlParser();
 
     @BeforeAll
-    static void setupAuthenticator() throws IOException {
+    public static void setupAuthenticator() throws IOException {
         Platform platform = Platform.getInstance();
         if (!platform.hasRoute("v1.api.auth")) {
             LambdaFunction f = (headers, input, instance) -> {