+++ system/platform-core/src/test/java/org/platformlambda/core/MulticastTest.java
@@ -41,7 +41,7 @@ class MulticastTest extends TestBase {
     private static final int WAIT_INTERVAL = 300;
 
     @BeforeAll
-    public static void setup() throws InterruptedException {
+    static void setup() throws InterruptedException {
         // The multicast.yaml configuration will be loaded when the EventEmitter singleton initializes
         EventEmitter po = EventEmitter.getInstance();
         log.info("Multicast ready? {}", po.isMulticastEnabled());